requirements.txt
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.1
apache-airflow==2.10.1
airflow-provider-great-expectations==0.2.9
apache-airflow-providers-postgres==5.12.0
fastapi==0.114.2
fastapi[standard]
flake8==7.1.0
streamlit==1.38.0
sqlalchemy<2.0
pydantic==2.9.2
toml==0.10.2
great-expectations==0.18.21
mistune==3.0.2
python-dotenv
uvicorn
databases